Add Scaffold Generator
